# Conflicts:
#	pom.xml
#	src/main/java/me/botsko/prism/ConfigBase.java
#	src/main/java/me/botsko/prism/Prism.java
#	src/main/java/me/botsko/prism/PrismConfig.java
#	src/main/java/me/botsko/prism/actionlibs/ActionMessage.java
#	src/main/java/me/botsko/prism/actionlibs/ActionRegistry.java
#	src/main/java/me/botsko/prism/actionlibs/ActionsQuery.java
#	src/main/java/me/botsko/prism/actionlibs/HandlerRegistry.java
#	src/main/java/me/botsko/prism/actionlibs/Ignore.java
#	src/main/java/me/botsko/prism/actionlibs/InternalAffairs.java
#	src/main/java/me/botsko/prism/actionlibs/QueryParameters.java
#	src/main/java/me/botsko/prism/actionlibs/QueryResult.java
#	src/main/java/me/botsko/prism/actionlibs/RecordingQueue.java
#	src/main/java/me/botsko/prism/actionlibs/RecordingTask.java
#	src/main/java/me/botsko/prism/actions/BlockAction.java
#	src/main/java/me/botsko/prism/actions/BlockChangeAction.java
#	src/main/java/me/botsko/prism/actions/EntityAction.java
#	src/main/java/me/botsko/prism/actions/GenericAction.java
#	src/main/java/me/botsko/prism/actions/Handler.java
#	src/main/java/me/botsko/prism/actions/HangingItemAction.java
#	src/main/java/me/botsko/prism/actions/ItemStackAction.java
#	src/main/java/me/botsko/prism/actions/SignAction.java
#	src/main/java/me/botsko/prism/appliers/Preview.java
#	src/main/java/me/botsko/prism/appliers/Rollback.java
#	src/main/java/me/botsko/prism/commandlibs/Executor.java
#	src/main/java/me/botsko/prism/commandlibs/Flag.java
#	src/main/java/me/botsko/prism/commandlibs/PreprocessArgs.java
#	src/main/java/me/botsko/prism/commands/AboutCommand.java
#	src/main/java/me/botsko/prism/commands/ActionsCommand.java
#	src/main/java/me/botsko/prism/commands/DeleteCommand.java
#	src/main/java/me/botsko/prism/commands/DrainCommand.java
#	src/main/java/me/botsko/prism/commands/ExtinguishCommand.java
#	src/main/java/me/botsko/prism/commands/LookupCommand.java
#	src/main/java/me/botsko/prism/commands/NearCommand.java
#	src/main/java/me/botsko/prism/commands/PageCommand.java
#	src/main/java/me/botsko/prism/commands/PreviewCommand.java
#	src/main/java/me/botsko/prism/commands/PrismCommands.java
#	src/main/java/me/botsko/prism/commands/ReportCommand.java
#	src/main/java/me/botsko/prism/commands/ResetmyCommand.java
#	src/main/java/me/botsko/prism/commands/RestoreCommand.java
#	src/main/java/me/botsko/prism/commands/RollbackCommand.java
#	src/main/java/me/botsko/prism/commands/SetmyCommand.java
#	src/main/java/me/botsko/prism/commands/TeleportCommand.java
#	src/main/java/me/botsko/prism/commands/UndoCommand.java
#	src/main/java/me/botsko/prism/commands/ViewCommand.java
#	src/main/java/me/botsko/prism/commands/WandCommand.java
#	src/main/java/me/botsko/prism/commands/WhatCommand.java
#	src/main/java/me/botsko/prism/database/QueryBuilder.java
#	src/main/java/me/botsko/prism/database/mysql/ActionReportQueryBuilder.java
#	src/main/java/me/botsko/prism/database/mysql/BlockReportQueryBuilder.java
#	src/main/java/me/botsko/prism/database/mysql/SelectQueryBuilder.java
#	src/main/java/me/botsko/prism/listeners/PrismBlockEvents.java
#	src/main/java/me/botsko/prism/listeners/PrismCustomEvents.java
#	src/main/java/me/botsko/prism/listeners/PrismEntityEvents.java
#	src/main/java/me/botsko/prism/listeners/PrismInventoryEvents.java
#	src/main/java/me/botsko/prism/listeners/PrismPlayerEvents.java
#	src/main/java/me/botsko/prism/listeners/PrismVehicleEvents.java
#	src/main/java/me/botsko/prism/listeners/PrismWorldEvents.java
#	src/main/java/me/botsko/prism/listeners/self/PrismMiscEvents.java
#	src/main/java/me/botsko/prism/measurement/Metrics.java
#	src/main/java/me/botsko/prism/measurement/QueueStats.java
#	src/main/java/me/botsko/prism/measurement/TimeTaken.java
#	src/main/java/me/botsko/prism/monitors/OreMonitor.java
#	src/main/java/me/botsko/prism/monitors/UseMonitor.java
#	src/main/java/me/botsko/prism/parameters/ActionParameter.java
#	src/main/java/me/botsko/prism/parameters/BlockParameter.java
#	src/main/java/me/botsko/prism/parameters/FlagParameter.java
#	src/main/java/me/botsko/prism/parameters/IdParameter.java
#	src/main/java/me/botsko/prism/parameters/PlayerParameter.java
#	src/main/java/me/botsko/prism/parameters/RadiusParameter.java
#	src/main/java/me/botsko/prism/parameters/SimplePrismParameterHandler.java
#	src/main/java/me/botsko/prism/players/PlayerIdentification.java
#	src/main/java/me/botsko/prism/purge/LogPurgeCallback.java
#	src/main/java/me/botsko/prism/purge/PurgeChunkingUtil.java
#	src/main/java/me/botsko/prism/purge/PurgeManager.java
#	src/main/java/me/botsko/prism/purge/SenderPurgeCallback.java
#	src/main/java/me/botsko/prism/utils/BlockUtils.java
#	src/main/java/me/botsko/prism/utils/DateUtil.java
#	src/main/java/me/botsko/prism/utils/ItemUtils.java
#	src/main/java/me/botsko/prism/utils/MiscUtils.java
#	src/main/java/me/botsko/prism/wands/InspectorWand.java
#	src/main/java/me/botsko/prism/wands/ProfileWand.java
#	src/main/java/me/botsko/prism/wands/RestoreWand.java
#	src/main/java/me/botsko/prism/wands/RollbackWand.java
#	src/main/java/me/botsko/prism/wands/Wand.java
#	src/main/java/me/botsko/prism/wands/WandBase.java
#	src/main/resources/items.yml
#	src/main/resources/plugin.yml

+++ src/main/java/me/botsko/prism/parameters/IdParameter.java
@@ -1,6 +1,6 @@
 package me.botsko.prism.parameters;
 
-import com.helion3.prism.libs.elixr.TypeUtils;
+import me.botsko.prism.utils.TypeUtils;
 import me.botsko.prism.actionlibs.QueryParameters;
 import org.bukkit.command.CommandSender;
 
@@ -8,21 +8,22 @@ import java.util.regex.Pattern;
 
 public class IdParameter extends SimplePrismParameterHandler {
 
-    /**
+	/**
 	 * 
 	 */
-    public IdParameter() {
-        super( "ID", Pattern.compile( "[\\d,]+" ), "id" );
-    }
+	public IdParameter() {
+		super("ID", Pattern.compile("[\\d,]+"), "id");
+	}
 
-    /**
+	/**
 	 * 
 	 */
-    @Override
-    public void process(QueryParameters query, String alias, String input, CommandSender sender) {
+	@Override
+	public void process(QueryParameters query, String alias, String input, CommandSender sender) {
 
-        if( !TypeUtils.isNumeric( input ) ) { throw new IllegalArgumentException(
-                "ID must be a number. Use /prism ? for help." ); }
-        query.setId( Integer.parseInt( input ) );
-    }
+		if (!TypeUtils.isNumeric(input)) {
+			throw new IllegalArgumentException("ID must be a number. Use /prism ? for help.");
+		}
+		query.setId(Long.parseLong(input));
+	}
 }
